surgery handling and suit stuff

+++ modular_zubbers/code/modules/customization/species/proteans/protean_modsuit.dm
@@ -4,23 +4,33 @@
 	theme = /datum/mod_theme // Standard theme. TODO: Can be changed with standard mod armors
 
 	applied_core = /obj/item/mod/core/protean
-	applied_cell = null
+	applied_cell = null // Goes off stomach
+
+	resistance_flags = LAVA_PROOF | FIRE_PROOF | UNACIDABLE | ACID_PROOF // funny nanite
+	/// Whether or not the wearer can undeploy parts.
+	var/modlocked = FALSE
 
 /datum/mod_theme/protean
 	name = "Protean"
 
 /obj/item/mod/control/pre_equipped/protean/Initialize(mapload, datum/mod_theme/new_theme, new_skin, obj/item/mod/core/new_core)
 	. = ..()
-	ADD_TRAIT(src, TRAIT_NODROP, QUIRK_TRAIT)
+	ADD_TRAIT(src, TRAIT_NODROP, "protean")
 	actions_types += list()
 
 /obj/item/mod/control/pre_equipped/protean/wrench_act(mob/living/user, obj/item/wrench)
 	return FALSE // Can't remove the core.
 
+/obj/item/mod/control/pre_equipped/protean/emag_act(mob/user, obj/item/card/emag/emag_card)
+	return FALSE // Nope
+
 /obj/item/mod/control/pre_equipped/protean/canStrip(mob/who)
 	return TRUE
 
 /obj/item/mod/control/pre_equipped/protean/doStrip(mob/stripper, mob/owner) // Custom stripping code.
+	if(!isprotean(wearer)) // Strip it normally
+		REMOVE_TRAIT(src, TRAIT_NODROP, "protean") // Your ass is coming off.
+		return ..()
 	var/obj/item/mod/module/storage/inventory = locate() in src.modules
 	if(!isnull(inventory))
 		src.atom_storage.remove_all()
@@ -37,3 +47,18 @@
 		wearer.dropItemToGround(src, TRUE, TRUE)
 	else
 		forceMove(get_turf(loc))
+
+/// Proteans can lock themselves on people.
+/obj/item/mod/control/pre_equipped/protean/proc/toggle_lock()
+	if(modlocked)
+		REMOVE_TRAIT(src, TRAIT_NODROP, "protean")
+	modlocked = !modlocked
+
+/obj/item/mod/control/pre_equipped/protean/equipped(mob/user, slot, initial)
+	. = ..()
+
+	if(!isprotean(wearer))
+		return
+	if(slot == ITEM_SLOT_BACK && modlocked)
+		ADD_TRAIT(src, TRAIT_NODROP, "protean")
+		to_chat(wearer, span_warning("The suit does not seem to be able to come off..."))
